Рефакториг, обьединил компоненты film-details и film-info
